refactor: move the common txn params to common.rs
