+++ crates/algokit_utils/src/transactions/mod.rs
@@ -1,7 +1,6 @@
+pub mod common;
 pub mod composer;
 
 // Re-export commonly used transaction types
-pub use composer::{
-    CommonParams, Composer, ComposerError, ComposerTxn, EmptySigner, PaymentParams, TxnSigner,
-    TxnSignerGetter,
-};
+pub use common::{CommonParams, DefaultSignerGetter, EmptySigner, TxnSigner, TxnSignerGetter};
+pub use composer::{Composer, ComposerError, ComposerTxn, PaymentParams};
